+++ src/index.test.ts
@@ -1425,6 +1425,17 @@ describe('combine', () => {
   });
 });
 
+test('cwise', () => {
+  const a = StridedView.of(A6, [2, 3]);
+  const b = StridedView.of(S6, [2, 3]);
+
+  expect(StridedView.cwise(a, b, (x, y) => x + y).toString()).toMatchInlineSnapshot(`
+    "1A,2B
+    3C,4D
+    5E,6F"
+  `);
+});
+
 describe('stress tests', () => {
   const sx = 30;
   const sy = 50;